+++ src/components/common/info/info-exp/tch-ip-exp/ip-item/IpItem.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import './ipItem.css'
-import { Icon, message } from 'antd'
+import { Icon, message, Modal } from 'antd'
 import moment from 'moment'
 import { UserInfoApi } from 'src/ajax'
 
@@ -10,15 +10,20 @@ export default class IpItem extends React.Component {
     this.props.setSelectedItem(exp)
   }
 
-  async handleDelete(exp) {
-    try {
-      await UserInfoApi.deleteIp(exp.id)
-      this.props.deleteItem(exp)
-      message.success('删除成功')
-    } catch (err) {
-      message.error('删除失败')
-      console.log(err)
-    }
+  handleDelete(exp) {
+    Modal.confirm({
+      title: '是否确认删除',
+      onOk: async () => {
+        try {
+          await UserInfoApi.deleteIp(exp.id)
+          this.props.deleteItem(exp)
+          message.success('删除成功')
+        } catch (err) {
+          message.error('删除失败')
+          console.log(err)
+        }
+      }
+    })
   }
 
   transfrmType(type) {
